+++ src/main/java/com/gwangya/lock/HazelcastLockService.java
@@ -8,7 +8,6 @@ import java.util.concurrent.TimeUnit;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-import com.gwangya.global.util.LockReleaseEventListener;
 import com.gwangya.lock.exception.LockOccupationException;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.cp.lock.FencedLock;
